improved virtual worksapce code
On-behalf-of: @SAP a.shcherbatiuk@sap.com
