+++ listener/reconciler/kcp/reconciler.go
@@ -19,6 +19,9 @@ type KCPReconciler struct {
 	log                        *logger.Logger
 	virtualWorkspaceReconciler *VirtualWorkspaceReconciler
 	configWatcher              *ConfigWatcher
+
+	// Components for controller setup (moved from constructor)
+	apiBindingReconciler *APIBindingReconciler
 }
 
 func NewKCPReconciler(
@@ -59,7 +62,7 @@ func NewKCPReconciler(
 		return nil, err
 	}
 
-	// Setup APIBinding reconciler
+	// Create APIBinding reconciler (but don't set up controller yet)
 	apiBindingReconciler := &APIBindingReconciler{
 		Client:              mgr.GetClient(),
 		Scheme:              opts.Scheme,
@@ -69,14 +72,6 @@ func NewKCPReconciler(
 		APISchemaResolver:   schemaResolver,
 		ClusterPathResolver: clusterPathResolver,
 		Log:                 log,
-	}
-
-	// Setup the controller with cluster context - this is crucial for req.ClusterName
-	if err := ctrl.NewControllerManagedBy(mgr).
-		For(&kcpapis.APIBinding{}).
-		Complete(kcpctrl.WithClusterInContext(apiBindingReconciler)); err != nil {
-		log.Error().Err(err).Msg("failed to setup APIBinding controller")
-		return nil, err
 	}
 
 	// Setup virtual workspace components
@@ -101,6 +96,7 @@ func NewKCPReconciler(
 		log:                        log,
 		virtualWorkspaceReconciler: virtualWorkspaceReconciler,
 		configWatcher:              configWatcher,
+		apiBindingReconciler:       apiBindingReconciler,
 	}, nil
 }
 
@@ -109,12 +105,30 @@ func (r *KCPReconciler) GetManager() ctrl.Manager {
 }
 
 func (r *KCPReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	// This method is not used - reconciliation is handled by the APIBinding controller
+	// This method is required by the reconciler.CustomReconciler interface but is not used directly.
+	// Actual reconciliation is handled by the APIBinding controller set up in SetupWithManager().
+	// KCPReconciler acts as a coordinator/manager rather than a direct reconciler.
 	return ctrl.Result{}, nil
 }
 
 func (r *KCPReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	// Controllers are already set up in the constructor
+	// Handle cases where the reconciler wasn't properly initialized (e.g., in tests)
+	if r.apiBindingReconciler == nil {
+		if r.log != nil {
+			r.log.Debug().Msg("APIBinding reconciler not initialized, skipping controller setup")
+		}
+		return nil
+	}
+
+	// Setup the APIBinding controller with cluster context - this is crucial for req.ClusterName
+	if err := ctrl.NewControllerManagedBy(mgr).
+		For(&kcpapis.APIBinding{}).
+		Complete(kcpctrl.WithClusterInContext(r.apiBindingReconciler)); err != nil {
+		r.log.Error().Err(err).Msg("failed to setup APIBinding controller")
+		return err
+	}
+
+	r.log.Info().Msg("Successfully set up APIBinding controller")
 	return nil
 }
